# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MultiTenantEcommerce API",
        "description": "هذا هو التوثيق الرسمي لـ MultiTenantEcommerce API",
        "contact": {
            "email": "support@yourdomain.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "الخادم الأساسي"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "تسجيل الدخول",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم إرسال رمز التحقق إلى البريد الإلكتروني",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Your verification code has been sent to your email"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "بيانات غير صحيحة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "التحقق من الرمز",
                "operationId": "86a3349a0306cb26904691a7211faf38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم التحقق بنجاح وتم إنشاء التوكن",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "تسجيل مستخدم جديد",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء المستخدم بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "إرجاع قائمة الطلبات",
                "description": "عرض قائمة الطلبات",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "بحث عن طلب بالاسم أو الوصف",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم استرجاع قائمة الطلبات بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "إنشاء طلب جديد",
                "description": "إنشاء طلب جديد",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء الطلب بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطأ في التحقق من البيانات"
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "إرجاع بيانات طلب معين",
                "description": "عرض تفاصيل طلب معين",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "رقم معرف الطلب",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم استرجاع بيانات الطلب بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "جلب قائمة المنتجات",
                "operationId": "76bf8683b79dbaa67be3fe1f4794c8f6",
                "responses": {
                    "200": {
                        "description": "قائمة المنتجات بنجاح"
                    },
                    "401": {
                        "description": "غير مصرح لك"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "إضافة منتج جديد",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 999.99
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A powerful laptop"
                                    },
                                    "tenant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء المنتج بنجاح"
                    },
                    "400": {
                        "description": "بيانات غير صحيحة"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "طلب تسجيل الدخول",
                "description": "البيانات المطلوبة لتسجيل الدخول",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "OrderStoreRequest": {
                "title": "Order Store Request",
                "required": [
                    "product_id",
                    "quantity",
                    "total",
                    "grand_total"
                ],
                "properties": {
                    "product_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 199.99
                    },
                    "discount": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "grand_total": {
                        "type": "number",
                        "format": "float",
                        "example": 189.99
                    }
                },
                "type": "object"
            },
            "ProductStoreRequest": {
                "required": [
                    "name",
                    "price",
                    "tenant_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Laptop"
                    },
                    "price": {
                        "type": "number",
                        "example": 999.99
                    },
                    "description": {
                        "type": "string",
                        "example": "High-end gaming laptop"
                    },
                    "tenant_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "طلب تسجيل مستخدم جديد",
                "description": "البيانات المطلوبة لإنشاء حساب جديد",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "VerifyRequest": {
                "title": "طلب التحقق من الرمز",
                "description": "البيانات المطلوبة للتحقق من رمز OTP",
                "required": [
                    "email",
                    "code"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "code": {
                        "type": "integer",
                        "example": 123456
                    }
                },
                "type": "object"
            },
            "OrderResource": {
                "title": "Order",
                "description": "تمثيل بيانات الطلب",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 199.99
                    },
                    "discount": {
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "grand_total": {
                        "type": "number",
                        "format": "float",
                        "example": 189.99
                    },
                    "status": {
                        "type": "string",
                        "example": "Pending"
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "title": "Product",
                "description": "تمثيل بيانات المنتج",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "هاتف ذكي"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 599.99
                    },
                    "description": {
                        "type": "string",
                        "example": "هاتف ذكي حديث مع كاميرا 64MP"
                    },
                    "tenant": {
                        "$ref": "#/components/schemas/TenantResource"
                    }
                },
                "type": "object"
            },
            "TenantResource": {
                "title": "Tenant",
                "description": "تمثيل بيانات Tenant",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "شركة ABC"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "info@abc.com"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "description": "تمثيل بيانات المستخدم",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-19T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-19T10:00:00Z"
                    },
                    "token": {
                        "type": "string",
                        "example": "1|abcdef1234567890"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "API لتسجيل الدخول والتسجيل والتحقق"
        },
        {
            "name": "Orders",
            "description": "إدارة الطلبات في النظام"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}